+++ jjava-distro/src/test/java/org/dflib/jjava/distro/KernelStartupIT.java
@@ -7,6 +7,8 @@ import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import java.util.Map;
+import static org.hamcrest.Matchers.matchesPattern;
 
 class KernelStartupIT extends ContainerizedKernelCase {
 
@@ -18,5 +20,18 @@ class KernelStartupIT extends ContainerizedKernelCase {
         assertEquals(0, snippetResult.getExitCode(), snippetResult.getStdout());
         assertThat(snippetResult.getStdout(), not(containsString("|")));
         assertThat(snippetResult.getStdout(), containsString("1001.0"));
+    }
+
+    @Test
+    void startUp_scriptRequiresClasspath() throws Exception {
+        Map<String, String> env = Map.of(
+                Env.JJAVA_CLASSPATH, TEST_CLASSPATH,
+                Env.JJAVA_STARTUP_SCRIPT, "var obj = new org.dflib.jjava.Dummy()"
+        );
+        String snippet = "\"hash = \" + obj.hashCode()";
+        Container.ExecResult snippetResult = executeInKernel(snippet, env);
+
+        assertThat(snippetResult.getStdout(), not(containsString("|")));
+        assertThat(snippetResult.getStdout(), matchesPattern("(?s).*hash = -?\\d+.*"));
     }
 }