+++ jjava-kernel/src/main/java/org/dflib/jjava/kernel/JavaKernel.java
@@ -16,7 +16,7 @@ import org.dflib.jjava.jupyter.kernel.comm.CommManager;
 import org.dflib.jjava.jupyter.kernel.display.DisplayData;
 import org.dflib.jjava.jupyter.kernel.display.Renderer;
 import org.dflib.jjava.jupyter.kernel.history.HistoryManager;
-import org.dflib.jjava.jupyter.kernel.magic.MagicParser;
+import org.dflib.jjava.jupyter.kernel.magic.MagicsResolver;
 import org.dflib.jjava.jupyter.kernel.magic.MagicTranspiler;
 import org.dflib.jjava.jupyter.kernel.magic.MagicsRegistry;
 import org.dflib.jjava.jupyter.kernel.util.CharPredicate;
@@ -72,7 +72,7 @@ public class JavaKernel extends BaseKernel {
             JupyterIO io,
             CommManager commManager,
             Renderer renderer,
-            MagicParser magicParser,
+            MagicsResolver magicsResolver,
             MagicsRegistry magicsRegistry,
             boolean extensionsEnabled,
             StringStyler errorStyler,
@@ -88,7 +88,7 @@ public class JavaKernel extends BaseKernel {
                 io,
                 commManager,
                 renderer,
-                magicParser,
+                magicsResolver,
                 magicsRegistry,
                 extensionsEnabled,
                 errorStyler);
@@ -231,7 +231,7 @@ public class JavaKernel extends BaseKernel {
      */
     @Override
     public Object evalRaw(String source) {
-        return evaluator.eval(magicParser.resolveMagics(source));
+        return evaluator.eval(magicsResolver.resolve(source));
     }
 
     @Override
@@ -385,7 +385,7 @@ public class JavaKernel extends BaseKernel {
                     buildJupyterIO(jupyterEncoding),
                     buildCommManager(),
                     buildRenderer(),
-                    buildMagicParser(magicTranspiler),
+                    buildMagicsResolver(magicTranspiler),
                     buildMagicsRegistry(),
                     buildExtensionsEnabled(),
                     buildErrorStyler(),