+++ RELEASE-NOTES.md
@@ -1,3 +1,8 @@
+## 1.0-a7
+
+* #67 JJAVA_STARTUP_SCRIPT can't start up spark
+* #102 Eval startup snippets explicitly
+
 ## 1.0-a6
 
 * #68 CompilationException doesn't provide any information about the underlying problem